Add files via upload
An e-commerce website wishes to find the lucky customer who will
be eligible for full value cash back. For this purpose, a number N is fed
to the system. It will return another number that is calculated by an
algorithm. In the algorithm, a sequence is generated, in which each
number is the sum of the two preceding numbers. Initially the
sequence will have two 1’s in it. The system will return the Nth number
from the generated sequence which is treated as the order ID. The
lucky customer will be the one who has placed that order.
Write an algorithm to help the website find the lucky customer.